+++ TeamCode/build.gradle
@@ -32,9 +32,9 @@ repositories {
 
 dependencies {
     implementation project(':FtcRobotController')
+
     implementation "com.acmerobotics.roadrunner:ftc:0.1.14"
     implementation "com.acmerobotics.roadrunner:core:1.0.0"
     implementation "com.acmerobotics.roadrunner:actions:1.0.0"
     implementation "com.acmerobotics.dashboard:dashboard:0.4.16"
-
 }
